diskops: Add new method to create a LUKS LVM PV
This will create an LVM2 Physical Volume on the LUKS Container. As such
we use it as we would the rest of the LVM2 process, but instead of creating
a new partition, we create it in the LUKS Container, which by this point
will already have been opened and accessible via the device mapper.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>