strategy: Implement the encryption option
When enabling encryption, we simply issue a few pre-steps for the already
implemented LVM2 path. Firstly, we create a LUKS Container, open it, and
then we create a physical volume on that opened LUKS Container.

Everything else still works the same way, creating /boot, and having the
Logical Volumes on the Volume Group, leading to a simpler and more maintainable
implementation.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>