postinstall: Port to clr-boot-manager
Much of the manual configuration butchering is removed, as clr-boot-manager
is now responsible for all the heavy lifting when it comes to the kernel
and boot loader management.

The installer continues (for now) to install the bootloader initially,
however for UEFI this job will become CBM's in the future. For GRUB we
continue to edit the GRUB_BACKGROUND dynamically so that we have some kind
of styling regardless of the boot topology.

The important thing is, we no longer are responsible for the target cmdline
used, nor do we copy and kernel blobs anymore.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>